+++ src/modules/bring-id-modal/component-types.tsx
@@ -1,11 +1,17 @@
-import { TGenerateSignature, TMode } from "@/types"
+import { TGenerateSignature } from "@/types"
+
+export type TCustomTitles = {
+  pointsTitle: string
+  pointsShortTitle: string
+  scoreTitle: string
+}
 
 export type TProps = {
   address?: string
   generateSignature?: TGenerateSignature
   connectUrl?: string
   iframeOnLoad?: () => void
-  mode?: TMode,
   highlightColor?: string
   theme?: 'light' | 'dark'
+  customTitles?: TCustomTitles
 }