+++ src/modules/bring-id-sdk/types/constructor-args/index.tsx
@@ -1,7 +1,8 @@
 import { TMode } from "@/types"
 
 type TConstructorArgs = {
-  mode: TMode
+  appId: string
+  mode?: TMode
 }
 
 export default TConstructorArgs