+++ src/modules/bring-id-sdk/types/verify-humanity/index.tsx
@@ -2,7 +2,8 @@ import { TSemaphoreProof } from "@/types"
 
 type TArgs = {
   minPoints?: number;
-  scope?: string;
+  contract?: string;
+  context?: number;
   message?: string;
 }