+++ src/types/credential-group-points-map.ts
@@ -1,1 +1,0 @@
-export type TCredentialGroupPointsMap = Map<string, number>