+++ src/types/response.tsx
@@ -3,7 +3,6 @@ import TErrorType from "./error-type";
 
 type TRequest = {
   type: TRequestType
-  requestId: string
   error?: TErrorType
   payload?: any
 }