+++ src/types/task-check.ts
@@ -1,5 +1,0 @@
-export type TTaskCheck = {
-  key: string
-  type: string
-  value: string
-}