+++ src/utils/index.ts
@@ -1,5 +1,3 @@
-import generateId from './generate-id'
-
 import shortenString from './shorten-string';
 import defineApiUrl from './define-api-url';
 import api from './api';
@@ -8,10 +6,8 @@ import msToTime from './ms-to-time';
 import isValidInboundMessage from './validate-inbound-message';
 import isValidOutboundMessage from './validate-outbound-message';
 import fetchRegistryConfig from './fetch-registry-config'
-import fetchTasksConfig from './fetch-tasks-config'
 
 export {
-  generateId,
   shortenString,
   msToTime,
   defineApiUrl,
@@ -19,6 +15,5 @@ export {
   createQueryString,
   isValidInboundMessage,
   isValidOutboundMessage,
-  fetchRegistryConfig,
-  fetchTasksConfig
+  fetchRegistryConfig
 }